Move toDiscordComponent to UI.ts
