+++ src/UI/Button.ts
@@ -1,11 +1,9 @@
 import {
   ButtonInteraction,
   MessageButtonOptions,
-  MessageButtonStyle,
-  MessageSelectMenuOptions,
+  MessageButtonStyle
 } from 'discord.js';
-import { SelectMenuHandler } from './DispatchSelectMenu';
-import { getID, UIComponent } from './UI';
+import {UIComponent} from './UI';
 
 export type ButtonHandler = (
   interaction: ButtonInteraction
@@ -20,13 +18,13 @@ export type LinkButtonOptions = MessageButtonOptions & {
   style: 'LINK';
 };
 
-function isLinkButtonOptions(
+export function isLinkButtonOptions(
   options: UIComponent
 ): options is LinkButtonOptions {
   return 'url' in options && 'style' in options && options.style === 'LINK';
 }
 
-function isRegularButtonOptions(
+export function isRegularButtonOptions(
   options: UIComponent
 ): options is ButtonOptions {
   return 'onClick' in options && 'style' in options;
@@ -39,31 +37,3 @@ function isSelectMenuOptions(
   return 'onSelect' in options;
 }
 */
-
-export function toDiscordComponent(
-  options: UIComponent,
-  buttonListeners: Map<string, ButtonHandler>,
-  selectMenuListeners: Map<string, SelectMenuHandler>
-): MessageButtonOptions | MessageSelectMenuOptions {
-  if (isLinkButtonOptions(options)) {
-    return {
-      ...options,
-      type: 'BUTTON',
-    };
-  } else if (isRegularButtonOptions(options)) {
-    // nonlink buttons must have a customId
-    const { onClick, ...buttonOptions } = options;
-    const id = getID(options.label ?? '<unlabeled>', 'button');
-    buttonListeners.set(id, onClick);
-    return { ...buttonOptions, type: 'BUTTON', customId: id };
-  } else {
-    const { onSelect, ...selectOptions } = options;
-    const id = getID(selectOptions.placeholder ?? '<noplaceholder>', 'select');
-    selectMenuListeners.set(id, onSelect);
-    return {
-      ...selectOptions,
-      type: 'SELECT_MENU',
-      customId: id,
-    };
-  }
-}
